+++ src/main/java/com/sevenb/recipes_manager/dto/SupplyLossOutputDTO.java
@@ -2,14 +2,14 @@ package com.sevenb.recipes_manager.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
+
 import java.time.LocalDate;
 
 
 @Data
 public class SupplyLossOutputDTO {
     private Long id;
-    private BigDecimal quantity;
+    private Double quantity;
     private String description;
     private LocalDate lossDate;
     private String SupplyName;
